refactor RawFileManager
